requirements.txt
## PhysGaussian requirements
h5py==3.10.0
numpy==1.24.1
opencv_python==4.8.1.78
opencv_python_headless==4.9.0.80
Pillow==10.2.0
plyfile==1.0.3
PyMCubes==0.1.4
pymeshlab==2023.12.post1
scipy==1.12.0
setuptools==68.0.0
taichi==1.5.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.66.1
warp_lang==0.10.1

## threestudio requirements
lightning==2.0.0
omegaconf==2.3.0
jaxtyping
typeguard
git+https://github.com/KAIR-BAIR/nerfacc.git@v0.5.2
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
diffusers<0.20
transformers==4.30.2
accelerate
opencv-python
tensorboard
matplotlib
imageio>=2.28.0
imageio[ffmpeg]
git+https://github.com/NVlabs/nvdiffrast.git
libigl
xatlas
trimesh[easy]
networkx
pysdf
wandb
gradio==4.11.0
git+https://github.com/ashawkey/envlight.git
torchmetrics
